✨ : worker window 생성
